+++ server/controllers/request.js
@@ -1,37 +1,13 @@
 const Request = require("../models/Request");
 const User = require("../models/User");
+const organizeRequests = require("../utils/organizeRequests");
 const asyncHandler = require("express-async-handler");
-
-const sortRequests = (listToSort) => {
-  return listToSort.sort(function(reqA, reqB) {
-    return Date.parse(reqA.duration.start) < Date.parse(reqB.duration.start);
-  });
-};
-const getNext = (upcomingRequests) => {
-  if (upcomingRequests.length > 0) {
-    nextRequest = upcomingRequests.reduce(function(reqA, reqB) {
-      return Date.parse(reqA.duration.start) <
-        Date.parse(reqB.duration.start) && reqA.accepted
-        ? reqA
-        : reqB;
-    });
-    if (!nextRequest.accepted) {
-      nextRequest = null;
-    }
-  }
-
-  return nextRequest;
-};
 
 // @route POST /request
 // @desc create a new pet sitting request
 // @access Private
-exports.makeRequest = asyncHandler(async (req, res) => {
+exports.makeRequest = asyncHandler(async (req, res, next) => {
   const body = req.body;
-  if (!(body || body.sitterId || body.ownerId || body.duration)) {
-    res.status(400);
-    throw new Error("Bad Request");
-  }
   const sitter = await User.findById(body.sitterId);
   if (!sitter) {
     res.status(404);
@@ -65,37 +41,11 @@ exports.makeRequest = asyncHandler(async (req, res) => {
 // @desc Get requests for a logged in user
 // @access Private
 exports.getRequest = asyncHandler(async (req, res, next) => {
-  const now = new Date();
-  let before = [];
-  let after = [];
-  let nextRequest = null;
   const requests = await Request.find({ sitterId: req.user.id });
-  for (let request of requests) {
-    request = JSON.parse(JSON.stringify(request));
-    const sitterInfo = await User.findById(request.sitterId);
-    request.sitterInfo = {};
-    request.sitterInfo.username = sitterInfo.username;
-    request.sitterInfo.email = sitterInfo.email;
-
-    if (Date.parse(now) < Date.parse(request.duration.start)) {
-      after.push(request);
-    } else {
-      before.push(request);
-    }
-  }
-
-  after = sortRequests(after);
-  before = sortRequests(before);
-  nextRequest = getNext(after);
-  if (nextRequest) {
-    after = after.filter((req) => {
-      return req._id != nextRequest._id;
-    });
-  }
-
+  const processedRequests = await organizeRequests(requests);
   res.status(200).json({
     success: {
-      requests: { after, before, nextRequest },
+      requests: processedRequests,
     },
   });
 });
@@ -104,26 +54,24 @@ exports.getRequest = asyncHandler(async (req, res, next) => {
 // @desc Update a user's request
 // @access Private
 exports.editRequest = asyncHandler(async (req, res, next) => {
-  const body = { ...req.params, ...req.query };
-  if (!(body || body.id || body.accepted)) {
-    res.status(400);
-    throw new Error("Bad Request");
-  }
-  const verifyUser = await Request.findOne({ _id: body.id });
+  const body = req.body;
+  const verifyUser = await Request.findOne({ _id: body._id });
 
   if (req.user.id != (verifyUser.ownerId || verifyUser.sitterId)) {
     res.status(401);
     throw new Error("Not authorized");
   }
   const updatedRequest = await Request.findOneAndUpdate(
-    { _id: body.id },
+    { _id: body._id },
     body,
     { new: true }
   );
   if (updatedRequest) {
+    const requests = await Request.find({ sitterId: req.user.id });
+    const processedRequests = await organizeRequests(requests);
     res.status(200).json({
       success: {
-        updatedRequest: updatedRequest,
+        updatedRequests: processedRequests,
       },
     });
   } else {
